requirements.txt
flask
flask-cors
mysql-connector-python
pip install flask-bcrypt pyjwt
flask-bcrypt
pyjwt